stager: Fix issue with init handling returns from signal handlers
This fixes an issue where the init process wasn't properly handling the
return from a signal handler. The pause(2) documentation clearly states
the pause call only returns when the process was interrupted by a signal
and the signal handler has finished running.

In this case, pause will verify it returned due to a signal handler and
re-run pause.

This also adds a loop to waitpid in the sigchld signal handler to
properly reap the pid.

Also updated the local configuration file used with `make run` to enable
debugging.

A follow up PR will be posted to fix the "run" call returning
immediately with a forked process, and at that time I'll add an
integration test around the behavior.