stager: Fix to ensure the run command returns when its process exits.
This fixes an issue where the run command would not return if the
process it called exited, but something else was still using its
stdout. For example, enter a container and run "sleep 100 &" to fork off
the sleep call, then exit, and the command would not return.

This was because the sleep call was still using the shell's
stdout. After the process exits, the 'run' handler was waiting for the
stdout copy goroutine to have finshed. However since another process was
using the stdout, the stdout handler was never closed.

To account for this, the run command will now wait up to 100ms for the
goroutine to finish before it will continue on.

This adds an integration test to verify the command didn't block. The
test also verifies #118 by ensuring the pod is still running after the
sleep would have expired.